fixed getCountBank - можно сделать запрос который вернет к-во записей - а не выгружать все и делать .size @PropertySource("classpath:application.properties") @PropertySource(value = "file:${prop.file}", ignoreResourceNotFound = true) @Controller public class CustomerController
@PropertySource - по-моему лишняя

DataCustomer - переделать

+++ src/main/java/com/dao/CustomerDataBase.java
@@ -10,5 +10,6 @@ public interface CustomerDataBase {
     Customer getCustomer(Customer customer);
     void updateCustomer(Customer customer);
     List<Customer> getAll();
-    int getCountBank(String sum, String period);
+    long getCountBank(String sum, String period);
+
 }
